Implement behaviour for raising PR if branch exists
TODO: this would need to be fixed up to the previous commit. I've
kept them separate for clarity.

NB: we _could_ take this further and check that the branch
contains the correct contents before we open the PR. I'd suggest
doing that in a future commit, by editing the existing test.