Make global and use-regex flags optional and add always-yes flag to bulk_replace
This commit introduces the following changes to the bulk-replace script:

The --global and --use-regex flags have been made optional by setting their required parameter to false and adding a default value of false for both flags.
A new flag, --always-yes, is added, which is not required and has a default value of false. When set to true, it automatically confirms all 'Continue on existing branch?' questions.
The confirm_action method now accepts an optional always_yes parameter that skips the confirmation prompt when set to true.
The bulk_replace method has been updated to include the always_yes parameter.
Some minor adjustments to the logic have been made, such as checking if a branch already exists before creating it and adding a variable to store if a PR already exists.
These changes make the script more flexible by allowing users to choose whether to replace all occurrences of content or use a regex pattern and to skip confirmation prompts when desired.