Implement CHANGELOG updater and supporting utils
This commit adds a new file changelog_updater.rb which implements a script to update CHANGELOG.md files in specified repositories with new changes, either as unreleased or by incrementing the version number based on a given change type (major, minor, or patch).

The changelog_updater function makes use of the following utility functions:

* confirm_action: Prompts the user for confirmation.
* semver_increment: Increments a given version string based on the change type.
* insert_unreleased_changes: Inserts new changes into the "Unreleased" section of a CHANGELOG.md file.
* insert_new_version_section: Inserts a new version section into the CHANGELOG.md file with the given changes.

The util.rb file is updated with additional utility functions to support the changelog updater:

* list_files_in_repo: Lists files in a repository.
* diff: Generates a diff between two strings. Moved from bulk_replace.rb.
* edit_content: Opens a temporary file in the user's preferred text editor to edit the content.

Additionally, error handling is added to the following functions in util.rb:

* create_branch!: Handle exceptions when creating a branch.
* commit_file!: Handle exceptions when committing a file.
* create_pr!: Handle exceptions when creating a PR.

This commit also includes the addition of the 'tempfile' dependency.